Additional classes: Our agency owns several other types of vehicles:
Hybrid plane - this is a vehicle that is both air, land and sea, motorized, that travels on a paved road. The aforementioned vehicle is for military use, the rest of the information will be provided during the construction of the object.
Please note that in order to "enable" this multiple authorization, the appropriate design pattern must be used. Also, carefully calculate where the aforementioned vehicles should be placed in the inheritance hierarchy.

Electric bicycle - a motorized land vehicle with a fuel consumption of 20 liters per unit of distance, and two wheels.